api/requirements.base.txt
Django>=4.0,<4.1
gunicorn==20.1.0
django-cors-headers==3.10.1
django-rest-framework
django-solo==2.0.0
djangorestframework-camel-case==1.3.0
Pillow>=9.4.0,<9.5
psycopg2-binary>=2.9,<2.10
